A simple changes in Customer.java, Main.java, Menu.java

+++ src/main/java/practice/test/neledva/Customer.java
@@ -7,13 +7,14 @@ class Customer {
     private String lastName;
     private int amountMoney;
 
-    public double getAmountMoney() {
+    public int getAmountMoney() {
         return amountMoney;
     }
 
     public int getIdCustomer() {
         return id;
     }
+
     protected String getFirstName() {
         return firstName;
     }
